requirements.txt
anthropic>=0.30.0
google-auth>=2.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.32.0
PyYAML>=6.0.0
pandas>=2.2.0
pytz>=2024.1
lxml>=5.0.0
